split logical interfaces
