+++ python/ray/data/_internal/logical/interfaces/logical_plan.py
@@ -1,0 +1,13 @@
+from .plan import Plan
+from .logical_operator import LogicalOperator
+
+class LogicalPlan(Plan):
+    """The plan with a DAG of logical operators."""
+
+    def __init__(self, dag: LogicalOperator):
+        self._dag = dag
+
+    @property
+    def dag(self) -> LogicalOperator:
+        """Get the DAG of logical operators."""
+        return self._dag
