+++ python/ray/data/_internal/logical/interfaces/plan.py
@@ -1,0 +1,12 @@
+from .operator import Operator
+
+class Plan:
+    """Abstract class for logical/physical execution plans.
+
+    This plan should hold an operator representing the plan DAG and any auxiliary data
+    that's useful for plan optimization or execution.
+    """
+
+    @property
+    def dag(self) -> Operator:
+        raise NotImplementedError
